Merge

+++ Final_MohaGee/src/main/java/com/kh/mohagee/showBoard/model/service/ShowBoardService.java
@@ -40,21 +40,24 @@ public class ShowBoardService {
 		  }
 		  
 		  
+		/*
+		 * String content = request.getParameter("bContent");
+		 * 
+		 * contents = contents.replace("\r\n","<br>");
+		 */
+
 		  // 2. 추가된 게시글의 번호를 가져와 // 첨부파일 추가하기 
 		  
 		  if(list.size() > 0) { 
 			  for(int i = 0; i < list.size(); i++) {
 				  ShowAttachment a = list.get(i);
 				  
+				  // 파일 레벨 0, 1 나누기
 				  if(i== 0) a.setbFileLevel(0);
 				  else  a.setbFileLevel(1);
 				  
-				  System.out.println("check : "+a);
 				 result = showBoardDAO.insertShowAttachment(a);
-		  
-				  System.out.println("result 확인 : " + result); // 확인용 result 값 변경하기 // 
-				  
-		  		// result = 0; 
+		  				  
 		  
 		  if(result < 1) 
 			  throw new ShowBoardException("첨부파일 추가 중 에러 발생!"); 
@@ -74,43 +77,45 @@ public class ShowBoardService {
 	}
 
 	// 게시글 수정하기
-	public int updateShowBoard(ShowBoard showBoard, List<ShowAttachment> list) {
-		
-		int result = 0;
+	   public int updateShowBoard(ShowBoard showBoard, List<ShowAttachment> list) {
+	      
+	      int result = 0;
 
-		List<ShowAttachment> originList = showBoardDAO.selectShowAttachment(showBoard.getbNo());
-		
-		result = showBoardDAO.updateShowBoard(showBoard);
+	      List<ShowAttachment> originList = showBoardDAO.selectShowAttachment(showBoard.getbNo());
+	      result = showBoardDAO.updateShowBoard(showBoard);
 
-		if (result > 0) {
-			if (originList.size() > 0) {
-				result = showBoardDAO.deleteShowAttachment(showBoard.getbNo());
-			}
+	      if (result > 0) {
+	         if (originList.size() > 0 && list.size() > 0) { // 원본도 있고, 변경하려는 이미지도 있을 때
+	            result = showBoardDAO.deleteShowAttachment(showBoard.getbNo());
+	            
+	            // i 가 0일 때 대표이미지를, 1이면 서브이미지를 구현하기 위한 대표 사진 번호를 i로 먼저 지정하기
+	            int i = 0; // 이렇게!
+	            
+	            for(ShowAttachment a : list) {
+	               
+	               if(i == 0) a.setbFileLevel(0); // 대표 이미지로 설정할려고 0을 set함 
+	                 else a.setbFileLevel(1);  // 서브 이미지로 설정할려고 1을 set함
+	               
+	               result = showBoardDAO.updateShowAttachment(a);
+	               
+	               i++; // i++ 로 대표와 서브 구분                
+	            }
+	         } else if ( list.size() > 0) { // 원본은 없고 변경 이미지만 있을 때
 
-			 if(list.size() > 0) { 
-				  for(int i = 0; i < list.size(); i++) {
-					  ShowAttachment a = list.get(i);
-					  
-					  if(i== 0) a.setbFileLevel(0);
-					  else  a.setbFileLevel(1);
-					  
-					  System.out.println("check : "+a);
-					 result = showBoardDAO.updateShowAttachment(a);
-			  
-					  System.out.println("result 확인 : " + result); // 확인용 result 값 변경하기 // 
-					  
-			  		// result = 0; 
-			  
-			  if(result < 1) 
-				  throw new ShowBoardException("첨부파일 추가 중 에러 발생!"); 
-			  } 
-		 }
-			 
-			return result;
-		}
-		 
-		return result;
-	}
+	            int i = 0; // 이렇게!
+	         
+	            for(ShowAttachment a : list) {
+	               if(i == 0) a.setbFileLevel(0); // 대표 이미지로 설정할려고 0을 set함 
+	                 else a.setbFileLevel(1);  // 서브 이미지로 설정할려고 1을 set함
+	               
+	               result = showBoardDAO.updateShowAttachment(a);
+	               
+	               i++; // i++ 로 대표와 서브 구분       
+	            }
+	         }
+	      }
+	      return result;
+	   }
 
 	// 게시글 삭제하기
 	public int deleteShowBoard(int bNo) {
